Implemented ranking functionality in results and enhanced UI
A significant update for improving the overall UI/UX of the website. First, integrated an additional field in the database to fetch the user's birth date. Also, added a position counter in the results array to display the ranking of the images.

In addition, the error messages and alerts have been revamped to be more user-centric and informative. Ensured the proper display of success/error messages using session variables.

Resized the viewing of the image in the portfolio for better user experience. Additionally, the messages for successful updates or errors are now displayed in modals, offering a sleek interface.

Implemented new style selections for uploaded images and ensured a smooth display for the gallery.

Overall, this commit includes improvements towards a more user-friendly interface, enhanced functionality and engaging user experience.